+++ changelog.md
@@ -5,7 +5,7 @@ Added Product, ProductList Cart components, Product and Cart services, Product m
 
 ## 20.07
 
-Updated project tot use Angilar 8.
+Updated project tot use Angular 8.
 Refactored Cart to show count and sum for added items. Added some minimal styling via component styling, added hover styling.